Add RISC-V support to crossdev
I've been working on a RISC-V Gentoo port, which lives in an overlay at
the moment.  Without this patch, when I "crossdev -t
riscv64-unknown-linux-gnu" I see

     * Host Portage ARCH:     amd64
     * Target Portage ARCH:   *

which means I don't get quite the right set of packages.

This patch adds the RISC-V tuples to crossdev's architecture mapping
list, which sets the correct target portage arch when building a RISC-V
cross environment.  With this patch, and an overlay that defines various
RISC-V ports, I can build a RISC-V cross environment on my amd64 host.
Note that I'm using crossdev-20151026 with this patch applied on my
machine.

Signed-off-by: Palmer Dabbelt <palmer@dabbelt.com>